+++ app/js/playlist/controllers/PlaylistCtrl.js
@@ -81,7 +81,7 @@ angular.module("FM.playlist.PlaylistCtrl", [
         $scope.page = {
             loading: false,
             pages: 1,
-            total: playlistData.meta.totalPages || 1
+            total: (Math.floor($scope.meta.total / 20) + 1)
         };
 
         /**
@@ -92,13 +92,13 @@ angular.module("FM.playlist.PlaylistCtrl", [
             PlayerQueueResource.query().$promise
                 .then(function (response){
                     $scope.playlist = response.items;
-                    $scope.page.total = response.meta.totalPages;
-                    $scope.page.pages = 1;
                 });
 
             PlayerQueueResource.meta().$promise
                 .then(function (response){
                     $scope.meta = response;
+                    $scope.page.total = (Math.floor(response.total / 20) + 1);
+                    $scope.page.pages = 1;
                 });
         };
 
@@ -116,7 +116,12 @@ angular.module("FM.playlist.PlaylistCtrl", [
                     $scope.playlist = $scope.playlist.concat(response.items);
 
                     $scope.page.loading = false;
-                    $scope.page.total = response.meta.totalPages ? response.meta.totalPages : 1;
+                });
+
+            PlayerQueueResource.meta().$promise
+                .then(function (response){
+                    $scope.meta = response;
+                    $scope.page.total = (Math.floor(response.total / 20) + 1);
                 });
         };